% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRNA-Tools-package.R
\docType{package}
\name{miRNA-Tools}
\alias{miRNA-Tools}
\alias{miRNA-Tools-package}
\title{miRNA-Tools: A package to manage miRNA sequencing data}
\description{
Description: This package contains multiple fonctions to allow the analysis of
   deep sequencing miRNA datasets.
}

