% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utr_count.R
\name{utr_count}
\alias{utr_count}
\title{Count number of occurence of DNA substring in UTR regions}
\usage{
utr_count(sequence_name, sequence, substring_length, genome, utr,
  reverse_complement = FALSE)
}
\arguments{
\item{sequence_name}{A \code{character} vector of length 1 corresponding to
name of the current sequence being analysed.}

\item{sequence}{A \code{character} vector or a \code{DNAString} object
corresponding to the miRNA sequence to analyze.}

\item{substring_length}{The length of each substring to use to find perfect
matches in the UTR regions.}

\item{utr}{Which UTR region to use? "5UTR", "3UTR" or "both".
Default: "5UTR".}

\item{reverse_complement}{Use the reverse complement of the sequence?
                          Default: FALSE.}
}
\value{
A \code{data.frame} with the following columns:
    \itemize{
        \item name The sequence_name argument value.
        \item sequence The sequence argument value.
        \item substring The current_substring.
        \item substring The position of the substring in the sequence.
        \item count The number of perfect matches for the substring in the
                    UTR regions.
        \item percent The percentage of all the possible substrings of
                      \code{substring_length} that have a smaller number of
                      perfect matches in the UTR regions.
        \item mirbase_count The number of perfect matches for the all the
                            miRBase substrings at the same position of its
                            respective sequence in the specified genome.
        \item mirbase_percent The percentage of all miRBase substrings at
                              the same position in its respective sequence
                              that have a smaller number of perfect matches
                              in the UTR region of the specified genome.
        \item reverse_complement The value of the reverse complement
                                 argument.
    }
}
\description{
This function will look for the number of occurence of every substrings of
length \code{substring_length} in the UTR regions of a given genome.
}

